+++ docs/changes.md
@@ -63,6 +63,7 @@ Changes in command line arguments:
 ### Tests
 - [NXDRIVE-1212](https://jira.nuxeo.com/browse/NXDRIVE-1212): Disable all server converters, enabled on-demand
 - [NXDRIVE-1246](https://jira.nuxeo.com/browse/NXDRIVE-1246): Fix pytest randombug plugin STRICT mode
+- [NXDRIVE-1281](https://jira.nuxeo.com/browse/NXDRIVE-1281): Cover the whole DirectEdit feature with tests
 - [NXDRIVE-1293](https://jira.nuxeo.com/browse/NXDRIVE-1293): Tweak timeouts in testing jobs on Jenkins
 - [NXDRIVE-1316](https://jira.nuxeo.com/browse/NXDRIVE-1316): Only run tests when code files are modified or build is launched manually